rename account details to personal details
