+++ client/src/helpers/Authentication/updatePersonalDetails.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 
-function updateAccountDetails(name, email) {
+function updatePersonalDetails(name, email) {
   return axios.put(`/user`, { name, email }, { withCredentials: true });
 }
 
-export default updateAccountDetails;
+export default updatePersonalDetails;
