+++ server/Routes/AuthRoute.js
@@ -4,7 +4,7 @@ const {
   Login,
   Logout,
   checkAuth,
-  updateAccountDetails,
+  updatePersonalDetails,
   updatePassword,
 } = require('../Controllers/AuthController');
 
@@ -12,7 +12,7 @@ router.get('/', checkAuth);
 router.post('/signup', Signup);
 router.post('/login', Login);
 router.post('/logout', Logout);
-router.put('/', updateAccountDetails);
+router.put('/', updatePersonalDetails);
 router.put('/password', updatePassword);
 
 module.exports = router;
